requirements.txt
fastapi
uvicorn
pyjwt
pwdlib[argon2]
sqlmodel
python-dotenv
